Add scripts, refactor install to be easier to add elements
Adds a script folder to PATH and links it contents. They are stupid,
because there is one enemy I hate the most: my future self.

Since my past self is an asshole I had to refactor install.sh so that it
is easier to add new linkables. It has the downside that you cannot just
link certain elements, but who cares.